uci: fix arm64 compiling
Fixes:
https://github.com/openwrt/packages/issues/12488#issuecomment-643138068

+++ src/utils/dawn_uci.c
@@ -41,7 +41,7 @@ void uci_get_hostname(char* hostname)
         {
             len = dot - ptr.o->v.string;
         }
-        snprintf(hostname, HOST_NAME_MAX, "%.*s", len, ptr.o->v.string);
+        snprintf(hostname, HOST_NAME_MAX, "%.*s", (int)len, ptr.o->v.string);
     }
 
     uci_free_context(c);
